+++ hypothesis-python/docs/changelog.rst
@@ -18,6 +18,36 @@ Hypothesis 6.x
 
     .. include:: ../RELEASE.rst
 
+.. _v6.138.11:
+
+---------------------
+6.138.11 - 2025-09-01
+---------------------
+
+This patch updates our vendored `list of top-level domains <https://www.iana.org/domains/root/db>`__,
+which is used by the provisional :func:`~hypothesis.provisional.domains` strategy.
+
+.. _v6.138.10:
+
+---------------------
+6.138.10 - 2025-08-31
+---------------------
+
+Internal refactor to simplify |SearchStrategy|.
+
+.. _v6.138.9:
+
+--------------------
+6.138.9 - 2025-08-31
+--------------------
+
+This patch further improves stringification of lambdas, by
+never returning a lambda source unless it is confirmed to
+compile to the same code object. This stricter check makes
+it possible to widen the search for a matching source block,
+so that it can often be found even if the file has been
+edited.
+
 .. _v6.138.8:
 
 --------------------